feat: PlayerScreen Components 분리
