+++ androidApp/src/main/java/com/devjeong/watermelon_player/android/player/presentation/components/CoverScreen.kt
@@ -1,2 +1,27 @@
 package com.devjeong.watermelon_player.android.player.presentation.components
 
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.width
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ContentImage(imageResId: Int) {
+    Image(
+        painter = painterResource(id = imageResId),
+        contentDescription = "Content Image",
+        modifier = Modifier
+            .shadow(
+                elevation = 10.dp,
+                spotColor = Color(0x26000000),
+                ambientColor = Color(0x26000000)
+            )
+            .width(344.dp)
+            .height(344.dp)
+    )
+}
